ExprSecItemStackCreate - switch back to standard expression - was causing issues with conditions
